multiplexer blocks at sender, not receiver
Before now, a multiplexer could become "clogged" if one of the multiplexed connections was not being read from. In other words, blocking would occur at the receiving end, not the sending end. This meant that every time you created a Listener or a Connection (using a multiplexer), you had to read from it in a loop until it was finished.

Now, blocking takes place at the sending end. If a sub-connection is not being read from, Send()s to it will block. This way, other multiplexed sub-connections are not jammed, because the socket is not receiving any packets that cause it to block.