Close() should unblock all send/receives
Before, Close() depended on the remote sending a CloseAck. If the remote did not send a CloseAck, then Close() would not unblock send/receive calls. This change allows things outside of the main read loop to close() a bin's channels, making it possible to shutdown the bin immediately in the Close() call.